Changed freelist to use u64

+++ engine/src/containers/freelist.c
@@ -4,15 +4,15 @@
 
 typedef struct freelist_node
 {
-    u32 offset;
-    u32 size;
+    u64 offset;
+    u64 size;
     struct freelist_node* next;
 } freelist_node;
 
 typedef struct internal_state
 {
-    u32 total_size;
-    u32 max_entries;
+    u64 total_size;
+    u64 max_entries;
     freelist_node* head;
     freelist_node* nodes;
 } internal_state;
@@ -20,10 +20,10 @@ typedef struct internal_state
 freelist_node* get_node(freelist* list);
 void return_node(freelist* list, freelist_node* node);
 
-void freelist_create(u32 total_size, u64* memory_requirement, void* memory, freelist* out_list)
+void freelist_create(u64 total_size, u64* memory_requirement, void* memory, freelist* out_list)
 {
     // Enough space to hold state, plus array for all nodes
-    u32 max_entries = (total_size / sizeof(void*));
+    u64 max_entries = (total_size / sizeof(void*));
     *memory_requirement = sizeof(internal_state) + (sizeof(freelist_node) * max_entries);
     if (!memory)
         return;
@@ -49,7 +49,7 @@ void freelist_create(u32 total_size, u64* memory_requirement, void* memory, free
 
     // Invalidate the offset and size for all but the first node.
     // Invalid value will be checked for when seeking a new node from the list
-    for (u32 i = 1; i < state->max_entries; ++i)
+    for (u64 i = 1; i < state->max_entries; ++i)
     {
         state->nodes[i].offset = INVALID_ID;
         state->nodes[i].size = INVALID_ID;
@@ -67,7 +67,7 @@ void freelist_destroy(freelist* list)
     }
 }
 
-b8 freelist_allocate_block(freelist* list, u32 size, u32* out_offset)
+b8 freelist_allocate_block(freelist* list, u64 size, u64* out_offset)
 {
     if (!list || !out_offset || !list->memory)
         return false;
@@ -114,7 +114,7 @@ b8 freelist_allocate_block(freelist* list, u32 size, u32* out_offset)
     return false;
 }
 
-b8 freelist_free_block(freelist* list, u32 size, u32 offset)
+b8 freelist_free_block(freelist* list, u64 size, u64 offset)
 {
     if (!list || !list->memory || !size)
         return false;
@@ -196,7 +196,7 @@ void freelist_clear(freelist* list)
     internal_state* state = list->memory;
     // Invalidate offset and size for all but the first node. The invalid
     // value will be checked for when seeking a new node from the list
-    for (u32 i = 1; i < state->max_entries; ++i)
+    for (u64 i = 1; i < state->max_entries; ++i)
     {
         state->nodes[i].offset = INVALID_ID;
         state->nodes[i].size = INVALID_ID;
@@ -228,7 +228,7 @@ u64 freelist_free_space(freelist* list)
 freelist_node* get_node(freelist* list)
 {
     internal_state* state = list->memory;
-    for (u32 i = 1; i < state->max_entries; ++i)
+    for (u64 i = 1; i < state->max_entries; ++i)
     {
         if (state->nodes[i].offset == INVALID_ID)
             return &state->nodes[i];
